sqldelight support for database

+++ shared/src/commonMain/kotlin/com/krossovochkin/kweather/shared/feature/weatherdetails/data/WeatherDetailsDto.kt
@@ -10,7 +10,7 @@ data class WeatherDetailsDto(
 
     @Serializable
     data class MainDto(
-        val temp: Int
+        val temp: Double
     )
 
     @Serializable
